requirements.txt
google-generativeai>=0.8.0
python-telegram-bot>=20.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.30
apscheduler>=3.10.0
